make shutdown futures more like 0.1 (#9)
This changes the `Shutdown` future returned by `Runtime::shutdown` and
`Runtime::shutdown_on_idle` to be a futures 0.1 `Future`, so that users
can `.wait()` on it. If an 0.2 future is needed, users can still call
`.compat()` on it, but there is no futures 0.2 as 0.1 compatibility
layer.

Note that the work of shutting down the runtime still happens
synchronously, rather than in the future. I didn't want to run the risk
of leaking tasks if the `Shutdown` future is not driven.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>