runtime: fix `shutdown_on_idle` completing early
Currently, `Runtime::shutdown_on_idle` and
`current_thread::Runtime::run` can finish immediately if a runtime has
ever run a future that counts towards idleness previously. This means
that any spawned futures are not run.

This branch fixes the problem by checking the number of spawned futures
when an idle notification is received. If it's non-zero, then the idle
notification is stale and came from a previous `block_on` call. In that
case, we skip it and wait for the next one.

I've also added tests for this issue (and confirmed they fail against
master).

Signed-off-by: Eliza Weisman <eliza@buoyant.io>