Fix possible fd leak with threaded runtime (#28)
This PR introduces a fix where creating many runtimes in a loop will
cause an fd leak. This is because storing a strong arc within each
`on_thread_start` fn will cause the runtime handle to not be deallocated
when the `compat::Runtime` is dropped. To avoid this we only store a
weak pointer within the `on_thread_start` fn and attempt to upgrade it
to get the handle. Since, `on_thread_start` is generally always called
while the tokio runtime is not dropped the upgrade should not fail. This
then allows us to drop the runtime and deallocate the handle by storing
the only long lived strong arc within `compat::Runtime`.

I've verified this fixes our repro and doesn't leak fds on our
benchmarks where we orignally discovered this issue.

Closes #27

Signed-off-by: Lucio Franco <luciofranco14@gmail.com>
Co-authored-by: Eliza Weisman <eliza@buoyant.io>